+++ angular/projects/lib/src/lib/gridstack-item.component.ts
@@ -1,5 +1,5 @@
 /**
- * gridstack-item.component.ts 10.3.1-dev
+ * gridstack-item.component.ts 11.0.0
  * Copyright (c) 2022-2024 Alain Dumesny - see GridStack root license
  */