+++ src/dd-element.ts
@@ -1,5 +1,5 @@
 /**
- * dd-elements.ts 10.3.1-dev
+ * dd-elements.ts 11.0.0
  * Copyright (c) 2021-2024 Alain Dumesny - see GridStack root license
  */