+++ src/gridstack.ts
@@ -1,5 +1,5 @@
 /*!
- * GridStack 10.3.1-dev
+ * GridStack 11.0.0
  * https://gridstackjs.com/
  *
  * Copyright (c) 2021-2024  Alain Dumesny
@@ -1920,7 +1920,7 @@ export class GridStack {
     return this;
   }
 
-  static GDRev = '10.3.1-dev';
+  static GDRev = '11.0.0';
 
   /* ===========================================================================================
    * drag&drop methods that used to be stubbed out and implemented in dd-gridstack.ts